requirements.txt
mcp[cli] @ git+https://github.com/modelcontextprotocol/python-sdk.git@main
sqlalchemy
psycopg2-binary
openai
python-dotenv
